+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Breaking changes:
 * Split off `Request.sendNoBodyResponse()` from `sendContent()`, and made the
   latter take separate `body` and `contentType` arguments instead of those being
   part of the `options`.
+* Changed `Request.protocolName` to actually be the protocol name (and not just
+  a guess based on the port number).
+* Changed `HostInfo` to do its port-defaulting based on a port number (instead
+  of a "protocol").
 
 Other notable changes:
 * Added support for sending and receiving cookies.